ingest/requirements.txt
PyMuPDF>=1.23.0
openai>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0

